+++ js/components/PhonesCatalog.js
@@ -1,7 +1,8 @@
-export default class PhonesCatalog {
+import Component from '../Component.js';
+
+export default class PhonesCatalog extends Component {
   constructor(element, props) {
-    this.element = element;
-    this.props = props;
+    super(element, props);
 
     this.render();
 
@@ -19,19 +20,6 @@ export default class PhonesCatalog {
     this.on('click', 'PhoneLink', (event) => {
       const phoneId = event.delegateTarget.dataset.phoneID;
       this.props.onPhoneSelected(phoneId);
-    });
-  }
-
-  on(eventName, elementName, callback) {
-    this.element.addEventListener(eventName, (event) => {
-      const delegateTarget = event.target.closest(`[data-element="${elementName}"]`);
-
-      if (!delegateTarget) {
-        return;
-      }
-
-      event.delegateTarget = delegateTarget;
-      callback(event);
     });
   }
